Message Class
Define 3 public constructors that take a byte array, a string, or a double as input. Each of these constructors calls a private constructor that takes all four fields as parameters.The private constructor is the one that actually initializes the fields.

parseDouble() method is a helper method that tries to convert a string to a double. If the conversion fails, it returns Double.NaN as specified.

Signed-off-by: Noam Rahat <noamrht@gmail.com>