Topic Class
`subs` is a list of agents that are subscribed to the topic, and `pubs` is a list of agents that can publish messages to the topic. The subscribe(), unsubscribe(), addPublisher(), and removePublisher() methods modify these lists as appropriate. The publish() method sends the given message to all subscribed agents by calling their callback() method.

Signed-off-by: Noam Rahat <noamrht@gmail.com>