TopicManagerSingleton Class
`TopicManager` is a static inner class of `TopicManagerSingleton`.
It has a private constructor and a private static final instance of itself, which is initialized when the class is loaded.
The `get()` method in `TopicManagerSingleton` returns this instance.

`TopicManager` holds a `ConcurrentHashMap` from strings to `Topic` instances.
The `getTopic()` method returns the `Topic` for a given name, creating it if it doesn't exist.
The `getTopics()` method returns a collection of all the `Topic` instances in the map.
The `clear()` method removes all entries from the map.

Signed-off-by: Noam Rahat <noamrht@gmail.com>